third_party/WebKit/LayoutTests/http/tests/inspector/network/network-cyrillic-xhr.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function test()
{
    function dumpRequest(request, callback)
    {
        if (!request)
            return callback();
        InspectorTest.addResult(request.url());

        function contentLoaded(content)
        {
            InspectorTest.addResult("resource.content: " + content);
            callback();
        }

        request.requestContent().then(contentLoaded);
    }

    InspectorTest.recordNetwork();
    InspectorTest.makeSimpleXHR("GET", "resources/cyrillic.html?sync", false, firstXHRLoaded);

    function firstXHRLoaded()
    {
        InspectorTest.makeSimpleXHR("GET", "resources/cyrillic.html?async", true, step2);
    }

    function step2()
    {
        var requests = InspectorTest.networkRequests();
        dumpRequest(requests[requests.length - 2], step3);
    }

    function step3()
    {
        var requests = InspectorTest.networkRequests();
        dumpRequest(requests[requests.length - 1], step4);
    }

    function step4()
    {
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p> Tests cyrillic xhr content is correctly loaded in inspector.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=79026">Bug 79026</a>
</body>
</html>
